Add amd64 assembly version of swapUint64()
I decided to have fun learning how to write assembly code in Go, and
this tiny library seemed like a good opportunity.
I originally wrote assembly versions for all three variants (uint16,
uint32 and uint64) but the assembly version of swapUint16() and
swapUint32() were slower than their Go counterparts. It seems this is
because the Go compiler can inline Go functions but not assembly ones.

As the only version that showed some improvements was the uint64 one, I
decided to only keep this one.

I added benchmark functions to be able to quantify the improvements,
both implementations can be run by specifying (or not) the tag "noasm":

	$ go test -bench=.
	goos: linux
	goarch: amd64
	pkg: github.com/tsuna/endian
	cpu: 13th Gen Intel(R) Core(TM) i9-13900
	Benchmark16_1000-32    	 2306198	       499.8 ns/op
	Benchmark32_1000-32    	 1634236	       731.6 ns/op
	Benchmark64_1000-32    	 1457139	       824.8 ns/op
	PASS
	ok  	github.com/tsuna/endian	3.557s

	$ go test -bench=. -tags noasm
	goos: linux
	goarch: amd64
	pkg: github.com/tsuna/endian
	cpu: 13th Gen Intel(R) Core(TM) i9-13900
	Benchmark16_1000-32    	 2263766	       498.6 ns/op
	Benchmark32_1000-32    	 1650574	       727.0 ns/op
	Benchmark64_1000-32    	  932718	      1267 ns/op
	PASS
	ok  	github.com/tsuna/endian	3.516s
