+++ uint64_little.go
@@ -3,20 +3,9 @@
 // Use of this source code is governed by the Apache License 2.0
 // that can be found in the COPYING file.
 
-// +build 386 amd64 amd64p32 arm arm64 ppc64le mipsle mips64le mips64p32le
+// +build 386 amd64,noasm amd64p32 arm arm64 ppc64le mipsle mips64le mips64p32le
 
 package endian
-
-// swapUint16 converts a uint16 to network byte order and back.
-func swapUint16(n uint16) uint16 {
-	return (n&0x00FF)<<8 | (n&0xFF00)>>8
-}
-
-// swapUint32 converts a uint16 to network byte order and back.
-func swapUint32(n uint32) uint32 {
-	return (n&0x000000FF)<<24 | (n&0x0000FF00)<<8 |
-		(n&0x00FF0000)>>8 | (n&0xFF000000)>>24
-}
 
 // swapUint64 converts a uint16 to network byte order and back.
 func swapUint64(n uint64) uint64 {
